// src/app/api/model/workingHoursDayEntry.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';


/**
 * Arbeitszeit-Konfiguration für einen einzelnen Wochentag
 */
export interface WorkingHoursDayEntry { 
    dayOfWeek: DayOfWeek;
    /**
     * true = Arbeitstag mit Zeitfenster, false = arbeitsfreier Tag
     */
    isWorkDay: boolean;
    /**
     * Arbeitsbeginn im Format HH:mm. Pflicht wenn isWorkDay = true.
     */
    startTime?: string;
    /**
     * Arbeitsende im Format HH:mm. Pflicht wenn isWorkDay = true. Muss nach startTime liegen.
     */
    endTime?: string;
}
export namespace WorkingHoursDayEntry {
}