// src/app/api/model/workingHoursResponse.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkingHoursDayEntry } from './workingHoursDayEntry';


/**
 * Wöchentliche Arbeitszeit-Vorlage des Nutzers
 */
export interface WorkingHoursResponse { 
    /**
     * Alle 7 Wochentage (Mo–So), jeder genau einmal
     */
    days: Array<WorkingHoursDayEntry>;
}